backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
supabase==2.7.2
google-generativeai==0.8.0
pydantic-settings==2.5.2
python-dotenv==1.0.1
python-multipart==0.0.9
Pillow==10.4.0
httpx==0.27.2
